LPS-44957 move to module
